refusal_direction/requirements.txt
torch>=2.1.0
transformers>=4.40.0
datasets>=2.18.0
huggingface_hub>=0.20.0
numpy>=1.24.0
pyyaml>=6.0
tqdm>=4.66.0
